Environmental compliance certificate upload

+++ app/Http/Controllers/StaffOne/StaffOneProjectController.php
@@ -46,6 +46,8 @@ class StaffOneProjectController extends Controller
 
     // Uploads
 
+    //buildingPermit
+
     public function buildingPermitTempUpload(Request $request){
         $request->validate([
             'building_permit' => ['required','mimes:pdf']
@@ -58,8 +60,6 @@ class StaffOneProjectController extends Controller
         $name = explode('/', $filePath);
         return $name[1];
     }
-
-
     public function buildingPermitRemoveUpload($fileName){
 
         // return $fileName;
@@ -71,7 +71,6 @@ class StaffOneProjectController extends Controller
             ], 200);
         }
     }
-
     public function buildingPermitReplaceUpload($id, $fileName){
         $data = Project::find($id);
         $oldFile = $data->building_permit;
@@ -96,15 +95,66 @@ class StaffOneProjectController extends Controller
             'status' => 'error'
         ], 200);
     }
+
+    //environmental_compliance_certificate
+
+    public function environmentalTempUpload(Request $request){
+        $request->validate([
+            'environmental_compliance_certificate' => ['required','mimes:pdf']
+        ]);
+        
+        $file = $request->environmental_compliance_certificate;
+        $fileGenerated = md5($file->getClientOriginalName() . time());
+        $fileName = $fileGenerated . '.' . $file->getClientOriginalExtension();
+        $filePath = $file->storeAs('temp', $fileName, 'public');
+        $name = explode('/', $filePath);
+        return $name[1];
+    }
+    public function environmentalRemoveUpload($fileName){
+
+        // return $fileName;
+        if (Storage::disk('public')->exists('temp/' . $fileName)) {
+            Storage::disk('public')->delete('temp/' . $fileName);
+
+            return response()->json([
+                'status' => 'remove'
+            ], 200);
+        }
+    }
+    public function environmentalReplaceUpload($id, $fileName){
+        $data = Project::find($id);
+        $oldFile = $data->environmental_compliance_certificate;
+
+        // return $oldFile;
+        $data->environmental_compliance_certificate = null;
+        $data->save();
+
+        if (Storage::disk('public')->exists('environmental_compliance_certificate/' . $oldFile)) {
+            Storage::disk('public')->delete('environmental_compliance_certificate/' . $oldFile);
+
+            if (Storage::disk('public')->exists('temp/' . $fileName)) {
+                Storage::disk('public')->delete('temp/' . $fileName);
+            }
+
+            return response()->json([
+                'status' => 'replace'
+            ], 200);
+        }
+
+        return response()->json([
+            'status' => 'error'
+        ], 200);
+    }
+
+
     
-
-
     public function store(ProjectStoreRequest $request)
     {
         // return $request;
 
         $data = $request->validated();
 
+        //building_permit
         if(!empty($data['building_permit']) && isset($request->building_permit[0]['response'])){
             $fileName = $request->building_permit[0]['response'];
             $data['building_permit'] = $fileName;
@@ -112,6 +162,18 @@ class StaffOneProjectController extends Controller
             if (Storage::disk('public')->exists('temp/' . $fileName)) {
                 // Move the file
                 Storage::disk('public')->move('temp/' . $fileName, 'building_permit/' . $fileName); 
+                Storage::disk('public')->delete('temp/' . $fileName);
+            }
+        }
+
+        //environmental_compliance_certificate
+        if(!empty($data['environmental_compliance_certificate']) && isset($request->environmental_compliance_certificate[0]['response'])){
+            $fileName = $request->environmental_compliance_certificate[0]['response'];
+            $data['environmental_compliance_certificate'] = $fileName;
+
+            if (Storage::disk('public')->exists('temp/' . $fileName)) {
+                // Move the file
+                Storage::disk('public')->move('temp/' . $fileName, 'environmental_compliance_certificate/' . $fileName); 
                 Storage::disk('public')->delete('temp/' . $fileName);
             }
         }
@@ -154,6 +216,19 @@ class StaffOneProjectController extends Controller
         } else {
             unset($data['building_permit']); 
         }
+        //environmental_compliance_certificate
+        if(!empty($data['environmental_compliance_certificate']) && isset($request->environmental_compliance_certificate[0]['response'])){
+            $fileName = $request->environmental_compliance_certificate[0]['response'];
+            $data['environmental_compliance_certificate'] = $fileName;
+
+            if (Storage::disk('public')->exists('temp/' . $fileName)) {
+                // Move the file
+                Storage::disk('public')->move('temp/' . $fileName, 'environmental_compliance_certificate/' . $fileName); 
+                Storage::disk('public')->delete('temp/' . $fileName);
+            }
+        } else {
+            unset($data['environmental_compliance_certificate']); 
+        }
 
         if($data['contractual'] === 0){
             $data['status'] = 'Material';
@@ -183,6 +258,12 @@ class StaffOneProjectController extends Controller
         if (!empty($project->building_permit)) {
             if (Storage::disk('public')->exists('building_permit/' . $project->building_permit)) {
                 Storage::disk('public')->delete('building_permit/' . $project->building_permit);
+            }
+        }
+        //environmental_compliance_certificate
+        if (!empty($project->environmental_compliance_certificate)) {
+            if (Storage::disk('public')->exists('environmental_compliance_certificate/' . $project->environmental_compliance_certificate)) {
+                Storage::disk('public')->delete('environmental_compliance_certificate/' . $project->environmental_compliance_certificate);
             }
         }
